Remove retired servers from list.
Remove the servers from the list for #21 - not sure how to get the seeds to add the missing servers at this time.  Will have to dig into that I suppose.  Although, I wonder if people actually use this feature.  If they do, wouldn't it make more sense to have a drop-down with a list of servers so they don't have to copy/paste or even better just a checkbox and autodetect server?